Add weather data fetch worker

+++ Rainmember/app/src/main/java/br/ufpe/cin/android/rainmember/data/WeatherApi.kt
@@ -2,7 +2,7 @@ package br.ufpe.cin.android.rainmember.data
 
 interface WeatherApi {
 
-    fun getCurrentWeather(latitude: Long, longitude: Long) : WeatherData
+    fun getCurrentWeather(latitude: Double, longitude: Double) : WeatherData
 
-    fun getWeatherForecast (latitude: Long, longitude: Long) : List<WeatherData>
+    fun getWeatherForecast (latitude: Double, longitude: Double) : List<WeatherData>
 }
